Moved 3K lines of form tests into separate files
