+++ test/integration/api/forms/delete-restore.js
@@ -1,0 +1,143 @@
+const { readFileSync } = require('fs');
+const appRoot = require('app-root-path');
+const should = require('should');
+const config = require('config');
+const superagent = require('superagent');
+const { DateTime } = require('luxon');
+const { testService } = require('../../setup');
+const testData = require('../../../data/xml');
+const { exhaust } = require(appRoot + '/lib/worker/worker');
+
+describe('api: /projects/:id/forms (delete, restore)', () => {
+
+  ////////////////////////////////////////////////////////////////////////////////
+  // DELETING AND RESTORING TESTING
+  ////////////////////////////////////////////////////////////////////////////////
+
+  describe('/:id DELETE', () => {
+    it('should reject unless the user can delete', testService((service) =>
+      service.login('chelsea', (asChelsea) =>
+        asChelsea.delete('/v1/projects/1/forms/simple').expect(403))));
+
+    it('should delete the form', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.get('/v1/projects/1/forms/simple')
+            .expect(404)))));
+
+    it('should log the action in the audit log', testService((service, { Projects, Forms, Users, Audits }) =>
+      service.login('alice', (asAlice) =>
+        Projects.getById(1).then((o) => o.get())
+          .then((project) => Forms.getByProjectAndXmlFormId(project.id, 'simple')).then((o) => o.get())
+          .then((form) => asAlice.delete('/v1/projects/1/forms/simple')
+            .expect(200)
+            .then(() => Promise.all([
+              Users.getByEmail('alice@getodk.org').then((o) => o.get()),
+              Audits.getLatestByAction('form.delete').then((o) => o.get())
+            ])
+            .then(([ alice, log ]) => {
+              log.actorId.should.equal(alice.actor.id);
+              log.acteeId.should.equal(form.acteeId);
+            }))))));
+
+    it('should delete all associated assignments', testService((service) =>
+      service.login('alice', (asAlice) =>
+        Promise.all([
+          asAlice.post('/v1/projects/1/app-users')
+            .send({ displayName: 'test app user' })
+            .expect(200)
+            .then(({ body }) => body.id),
+          asAlice.get('/v1/roles/app-user')
+            .expect(200)
+            .then(({ body }) => body.id)
+        ])
+          .then(([ fkId, roleId ]) => asAlice.post(`/v1/projects/1/forms/simple/assignments/${roleId}/${fkId}`)
+            .expect(200)
+            .then(() => asAlice.delete('/v1/projects/1/forms/simple')
+              .expect(200))
+            .then(() => asAlice.get('/v1/projects/1/assignments/forms/')
+              .expect(200)
+              .then(({ body }) => {
+                body.should.eql([]);
+              }))))));
+  });
+
+  describe('/:id/restore (undeleting trashed forms)', () => {
+    it('should reject restoring unless the user can delete', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => service.login('chelsea', (asChelsea) =>
+            asChelsea.post('/v1/projects/1/forms/1/restore').expect(403))))));
+
+    it('should reject restoring if referenced by the xml form id', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.post('/v1/projects/1/forms/simple/restore')
+            .expect(400)
+            .then(({ body }) => {
+              body.code.should.equal(400.11); // Invalid input data type
+            })))));
+
+    it('should restore a soft-deleted form', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.post('/v1/projects/1/forms/1/restore')
+            .expect(200))
+          .then(() => asAlice.get('/v1/projects/1/forms/simple')
+            .expect(200)))));
+
+    it('should log form.restore in audit log', testService((service, { Audits, Forms, Users }) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.post('/v1/projects/1/forms/1/restore')
+            .expect(200))
+          .then(() => Promise.all([
+            Users.getByEmail('alice@getodk.org').then((o) => o.get()),
+            Forms.getByProjectAndXmlFormId(1, 'simple').then((o) => o.get()),
+            Audits.getLatestByAction('form.restore').then((o) => o.get())
+          ])
+          .then(([ alice, form, log ]) => {
+            log.actorId.should.equal(alice.actor.id);
+            log.acteeId.should.equal(form.acteeId);
+          })))));
+
+    it('should restore a specific form by numeric id when multiple trashed forms share the same xmlFormId', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.post('/v1/projects/1/forms/')
+            .send(testData.forms.simple.replace('id="simple"', 'id="simple" version="two"'))
+            .set('Content-Type', 'application/xml')
+            .expect(200))
+          .then(() => asAlice.delete('/v1/projects/1/forms/simple')
+            .expect(200))
+          .then(() => asAlice.post('/v1/projects/1/forms/1/restore')
+            .expect(200)))));
+
+    it('should fail restoring a form that is not deleted', testService((service) =>
+      service.login('alice', (asAlice) =>
+        asAlice.post('/v1/projects/1/forms/1/restore')
+          .expect(404))));
+
+    it('should fail to restore a form when another active form with the same form id exists', testService((service, { Audits }) =>
+      service.login('alice', (asAlice) =>
+        asAlice.delete('/v1/projects/1/forms/simple')
+          .expect(200)
+          .then(() => asAlice.post('/v1/projects/1/forms/')
+            .send(testData.forms.simple.replace('id="simple"', 'id="simple" version="two"'))
+            .set('Content-Type', 'application/xml')
+            .expect(200))
+          .then(() => asAlice.post('/v1/projects/1/forms/1/restore')
+            .expect(409)
+            .then(({ body }) => {
+              body.code.should.equal(409.3);
+              body.details.fields.should.eql([ 'projectId', 'xmlFormId' ]);
+              body.details.values.should.eql([ '1', 'simple' ]);
+            })))));
+  });
+});
